+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This changelog follows the patterns described here: https://keepachangelog.com/e
 Subheadings to categorize changes are `added, changed, deprecated, removed, fixed, security`.
 
 ## Unreleased
+### added
+- Add bootstrap example for styling solution in examples/yew-bootstrap
+
+## Unreleased
 ### changed
 - Remove HTML glob in tailwind.config.js